HMF refactoring 1

+++ Source/Engine/Rendering/CommandRecorder.cpp
@@ -55,7 +55,7 @@ void TCommandRecorder<RenderAllocator>::Execute(CommandBuffer* commandBuffer)
         CmdBase* cmdDataPtr = reinterpret_cast<CmdBase*>(m_buffer.Data() + header.offset);
 
         AssertDebug(header.offset < m_buffer.Size(), "Header has offset {} which is greater or equal to the buffer's size ({})",
-                    header.offset, m_buffer.Size());
+                    uint64(header.offset), m_buffer.Size());
 
         if (!header.IsCustom())
         {
